lib/rlist: Introduce rlist cpu range iterator
This is inspired by the iov_iter system which provides a uniform iteration
API that can access different underlying storage providers.

The iteration language is "rlist cpu", which is the rlist-style range
description with type safety for CPU memory information.

Two kinds of entries can be returned by the iterator:
  1) A folio, starting offset and length, spanning multiple contiguous
     folios
  2) A physical address and a p2pdma_provider indicating who is the
     owner

The iterator design follows that of the rlist design in the prior patch, a
state for reading and a different kind of state for appending. This is
because the append state consumes more memory.

Currently 'struct rlist', 'struct bio' and a 'struct page *' array are
implemented as storage options. I expect there will be interest in
providing a scatterlist based option as well.

The purpose of bringing everything to a common iterator is to allow a new
implementation in the DMA API and GUP to work with multiple data types
without any data copying. Eg a rlist_cpu enabled GUP can continue to
populate a page * array output with the same code base.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>