iommu/dma: Implement native rlist dma_map_ops
The native implementation is modeled after the current dma-iommu.c
approach to IOVA allocation.

To allow PPC to re-use this code the primary algorithm itself is pushed
into the generic code and is presented to the driver as two iterator APIs
that spit out the IOVA to physical ranges that the IOMMU should map.

The generic code also computes the number of bytes of IOVA required, and
necessary alignment. The same IOVA allocator is used as the map_sg path.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>