PCI/P2PDMA: Move the DMA API helpers to p2pdma_provider
The current P2PDMA helper is tied to both the pgmap and to scatterlist.
Provide a more generic family of helpers that can be used in non-struct
page and non-scatterlist scenarios.

The main worker becomes:

int p2pdma_provider_map(struct device *consumer,
			struct p2pdma_provider *provider, phys_addr_t base,
			dma_addr_t *dma_out,
			struct p2pdma_provider_map_cache *cache);

And several wrappers are provided to allow working with struct page and sg
cases.

Update the dma_map_ops implementations to use the new wrappers.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>