mm/gup: Create a wrapper for pin_user_pages to return a rlist
pin_user_pages_rlist() reads the page table out of a mm and returns
a range as an rlist_cpu. rlist_cpu can store actual ranges of pages
so this allows the output to take advantage of high order pages without
expanding everything to struct page lists.

This is an interm step of just providing the API with the call signature
we desire. The trivial implementation is taken from ib_umem_get().

With future work the lower levels of GUP would directly call the rlsa_*
APIs to avoid the slow step through the page_list when high order pages
are present.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>