iommu/viot: Add iommu_viot_get_single_iommu()
This is the ACPI VIOT version like iommu_of_get_single_iommu().

It parses the ACPI table, confirms all entries points to a single
iommu_driver and then returns a pointer to it.

Also cache the u32 id list in the iommu_probe_info and provide a getter
function which re-parses in case we overflow the cache.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>