arm/io: Use "Qo" constraints for writeb/w/l/q()
ARM currently uses the "r" constraint and causes the compiler to put the
address in a single register and then uses a simple effective address:

	asm volatile("str %x0, [%x1]" : : "rZ" (val), "r" (addr));

But the actual HW can have a more complex effective address. Like x86 does
we can use a memory reference constraint and have the compiler generate
the effective address:

	asm volatile("str %x0, %1"
		     :
		     : "rZ"(val), "Qo"(*(volatile u64 *)(addr)));

The difference in assembly (gcc 13) is visible when consecutive stores are
codegened. For instance a copy of two qwords previously was:

        ldr     x3, [x1]
        str x3, [x0]
        ldr     x3, [x1, 8]
        add     x4, x0, 8
        str x3, [x4]

And with this change it can become:

        ldp     x10, x9, [x1]
        str x10, [x0]
        str x9, [x0, 8]

"Qo" is chosen as the constaint for a few reasons:

 - "m" is rejected because the virtualization stacks are not able to
   handle pre/post indexed forms that write to the index register like:
     str x1, [x0, 8]!
   And "m" will allow the compiler to emit that.

 - o selects only base register plus offset

 - Q allows base register only (no offset) on some compilers that
   won't allow o for 0 offset. gcc 13 doesn't seem to need this

Further the read side has the same basic issue, but it is left at 'r'
because it relies on alternates and the LDR/LDAR instructions have
different restrictions on what effective addresses are allowed.

bloat-o-meter on a arm64 kconfig brings notable improvement:

  add/remove: 3/5 grow/shrink: 23/206 up/down: 132/-1840 (-1708)

Primarily in drivers, the best improvements are functions setting several
registers off a common base. vnic_cq_init() is quite a simple function
that shows the improvement.

Note that clang sees no improvement as of 17.0.6.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>