iommupt: Add the x86 PAE page table format
This is used by x86 CPUs and can be used in both x86 IOMMUs. When the x86
IOMMU is running SVA it is using this page table format.

This implementation follows the AMD v2 io-pgtable version.

There is nothing remarkable here, the format has a variable top and
limited support for different page sizes and no contiguous pages support.

In principle this can support the 32 bit configuration with fewer table
levels.

FIXME: Compare the bits against the VT-D version too.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>