iommupt: Add the 32 bit ARMv7s page table format
ARMv7s is the 32 bit ARM page table format, it has only two levels. The
format behavior is good, but there are several additional complexities to
support this format, 32 bit DMA and the 1k allocations for the table
memory that are not ready.

It is an interesting demonstration of how a format can work when there are
many different bit encodings for the PTEs depending on page size and
level.

Also it translates the full 32 bit VA, and has a u32 VA with a u64 OA,
which excercies some unique corner cases.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>