iommu: Bind the iommu page table operations to iommupt
This makes iommupt a first class object in the iommu core, by placing a
pointer to it in the struct iommu_domain.

Since the core code knows about this it can directly call the API without
having to bounce through the domain ops. This eliminates a function
pointer indirection.

Looking forward this approach will also allow adjusting the iommupt ops
interface without having to go and edit every driver to bounce through the
domain ops.

Provide kconfigs to indicate what set of ops the iommu driver wants to use
and #ifdef out the domain versions if no driver wants them. This makes it
easier to be confident that there are no hidden places expecting these
interfaces. FIXME we may not want to do merge this but it makes my life
easier and is not hard to undo.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>