vfio/type1: Remove Fine Grained Superpages detection
VFIO is looking to enable an optimization where it can rely on the
unmap operation not splitting and returning the size of a larger IOPTE.

Many iommu drivers already work this way, and iommu_pt implementations
always do.

Further, it prefers to only do this if the iommu supports small contiguous
page sizes as unmapping page by page will be worse than calling
iommu_iova_to_phys().

Rework this to simply check for iommu_pt (ensure required unmap behavior)
and for 2*PAGE_SIZE in the page size bitmap.

AMDv1 was the only domain to trigger this behavior and it is now replaced
with iommu_pt, so this is equivalent behavior.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>