iommu/arm-smmu-v3: Put iopf enablement in the domain attach path
The API is changing to be domain focused and IOMMU_DEV_FEAT_IOPF/SVA will
be deleted.

Instead, a driver should enable iopf support when a domain with an
iopf_handler is attached, and disable iopf support when the domain is
removed.

Support for SVA does not need to be enabled, creating or attaching a SVA
domain will simply fail if the HW can't support it.

Move all the logic for controlling the iopf queue under
arm_smmu_attach_prepare(). Keep track of the number of domains on the
master (over all the SSIDs) that require iopf. When the last domain
requiring iopf is detacted destroy the the iopf queue.

Remove the sva_lock, this is all protected by the group mutex.

This fixes page fault support for ARM in iommufd as iommufd does not set
IOMMU_DEV_FEAT_SVA.

This also generalizes iopf support to work on all PAGING domains, not just
for SVA.

Reviewed-by: Jason Gunthorpe <jgg@nvidia.com>