cover-letter: Remove IOMMU_DEV_FEAT_SVA
The new method for driver fault reporting support relies on the domain
to specify a iopf_handler. The driver should detect this and setup the
HW when fault capable domains are attached.

Move SMMUv3 to use this method and have VT-D validate support during
attach so that all three fault capable drivers have a no-op FEAT_SVA.
Then remove it.

A followup series is needed to remove FEAT_IOPF, VT-D is still using it,
but it is a no-op on SMMUv3 and AMD now.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>