iommu/amd: Don't call report_iommu_fault()
This old style API is only used by drivers/gpu/drm/msm,
drivers/remoteproc/omap_remoteproc.c, and
drivers/remoteproc/qcom_q6v5_adsp.c none are used on x86 HW.

Remove the dead code to discourage new users.

Also remove the domain == NULL print because it was intended to protect
against a NULL deref inside report_iommu_fault() which is no longer
possible.

Just always print the fault in the same format if it could get a dev_data.
There is no value to be gained by also printing if the domain is NULL. In
today's kernel when the dev_data is populated the domain will be made
!NULL very quickly during iommu device probing.

Reviewed-by: Vasant Hegde <vasant.hegde@amd.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>