container_of: Add alloc_container()
It is quite common in Linux for subsystems to use a container_of()
relationship between the subsystem struct and the driver struct:

  struct subsystem {...};
  struct driver {struct subsystem subsys; ...}

A useful pattern to allocate and initialize this memory has the subsystem
allocate the memory, initialize struct subsystem and then return it back
to the driver for additional initialization:

  struct driver *drv = subsystem_alloc(sizeof(struct driver), ...)
  subsystem_put(&drv->subsys);

Some cases might callback to the driver using a provided ops to initialize
the struct driver.

This is intended to provide a simpler and more consistent API for
drivers. A driver never sees a partially initialized struct subsystem, and
always uses subsystem_put(). If init/deinit ops are used then the driver
has a cleanly paired setup/destroy function that is fully sequenced by the
subsystem including for all error cases.

To make this trivial the struct driver must place struct subsystem at the
start of the struct so offsetof() == 0 and thus container_of() is simply a
type case.

Add an alloc_container() helper that documents this pattern and provides
the required type checks:

   struct driver *drv;
   drv = alloc_container(struct driver, subsys, subsystem_alloc, ...);

The macro checks that offsetof(struct driver, subsys) == 0 and that
subsystem_alloc() returns the same type as &struct driver::subsys. It
transparently flows through an ERR_PTR or NULL generated by the allocation
function.

The intention is for the subsystem to wrap this in its own helper, eg:

 #define vfio_alloc_device(dev_struct, member, dev, ops) \
	alloc_container(struct dev_struct, member, _vfio_alloc_device, dev, ops)

Suggested-by: Dan Williams <dan.j.williams@intel.com>
Link: https://lore.kernel.org/linux-rdma/67a697f02bc0b_2d1e2948a@dwillia2-xfh.jf.intel.com.notmuch/
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>